Clean up the select shader system a bit.
Added an object holder to the select program in the scene object,
and an access function for this.
Added an access function for the select program in the visualizers as
well.

Camera and pick colour is no longer passed through the localSelect()
functions of the SceneObject.
This is now set in the base select() function in the SceneObject.

Updated select functions throughout.