ScenGraph module renamed. TriangleFacets render bug fix.
Fixed an IBO initialization bug for the TrianglesIBO
used by the TriangleFacetsDefaultVisualizer.

TF code moved to the TriangleFacetsVisualizer class,
which now is equal to the ways it is done for PCurves and PSurfs.

Two-way dependency is removed between the OpenGL and TriangleFacets
modules.
TriangleFacets now depends on OpenGL but not the other way around.

SceneGraph module is renamed to Scene, this is a more covering umbrella.